refactor(chatui): move the chat UI out of the SDK examples into its own module
It was never an example. It is an application: a product-shaped server with
its own HTTP API, its own frontend, and its own operational concerns, which
happens to be built on the SDK. Leaving it under sdk/go/examples implied
readers should study it to learn the SDK, and implied the SDK owns it.

chatui is now its own module, so what an SDK consumer actually depends on
stays visible: the SDK requires golang.org/x/net and nothing else, while the
app's dependencies are the app's. It reaches the SDK through a replace
directive until the SDK is tagged.

Split by responsibility rather than left as one file: main.go wires flags and
routes, server.go holds the conversation registry and handlers, events.go
flattens SDK events for the browser.

Verified after the move against a live daemon: the page serves, a conversation
resumes by ID across a restart of this server, and the agent still has the
context its earlier turns built.
