+++ chatui/.gitignore
@@ -1,0 +1,2 @@
+# Built by "go build ." in this directory.
+/chatui
